DozeSettings: Refactor and link with the lift-to-wake toggles from System>Gestures.
* Also, changed doze_pick_up_gesture_ambient to enabled by default. This is to address an edge case where the device doesnt have a pick up sensor, has pulse on notifications on, they couldnt adjust the vibration intensity due to it depending on doze_pick_up_gesture_ambient being enabled. This way, Even if the switch isnt visible to the user, its in the enabled state which allows changing the vibration intensity but doesn't have any negitive effect since they dont have a pick up sensor for wake, so its a No-Op.

To be honest, this code needs cleanup. The settings are spread all throughout, in some cases , duplicated and not linked with each other. (e.g. I change Tap to Wake to off in Display>Lock Screen but Tap to wake in SigmaSettings is still on instead of changing with the Display>Lock Screen. This is at least an improvement.

Signed-off-by: Matt Filetto <matt.filetto@gmail.com>